+++ database/factories/FieldResponseFactory.php
@@ -1,15 +1,13 @@
 <?php
 
-namespace Database\Factories;
+namespace LaraZeus\Bolt\Database\Factories;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
+use LaraZeus\Bolt\Models\FieldResponse;
 
 class FieldResponseFactory extends Factory
 {
-    public function getModel(): string
-    {
-        return config('zeus-bolt.models.FieldResponse');
-    }
+    protected $model = FieldResponse::class;
 
     /**
      * Define the model's default state.