+++ src/Models/Field.php
@@ -2,13 +2,14 @@
 
 namespace LaraZeus\Bolt\Models;
 
-use Database\Factories\FieldFactory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Database\Eloquent\Relations\HasOne;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use LaraZeus\Bolt\Database\Factories\FieldFactory;
 use Spatie\Translatable\HasTranslations;
 
 /**
@@ -31,7 +32,22 @@ class Field extends Model
         'options' => 'array',
     ];
 
-    protected static function newFactory(): FieldFactory
+    protected static function booted(): void
+    {
+        static::deleting(function (Field $field) {
+            if ($field->isForceDeleting()) {
+                $field->fieldResponses()->withTrashed()->get()->each(function ($item) {
+                    $item->forceDelete();
+                });
+            } else {
+                $field->fieldResponses->each(function ($item) {
+                    $item->delete();
+                });
+            }
+        });
+    }
+
+    protected static function newFactory(): Factory
     {
         return FieldFactory::new();
     }
@@ -48,9 +64,9 @@ class Field extends Model
         return $this->belongsToMany(config('zeus-bolt.models.Section'));
     }
 
-    /** @return HasOne<FieldResponse> */
-    public function fieldResponses(): HasOne
+    /** @return HasMany<FieldResponse> */
+    public function fieldResponses(): HasMany
     {
-        return $this->hasOne(config('zeus-bolt.models.FieldResponse'));
+        return $this->hasMany(config('zeus-bolt.models.FieldResponse'));
     }
 }