+++ src/Models/Section.php
@@ -2,12 +2,13 @@
 
 namespace LaraZeus\Bolt\Models;
 
-use Database\Factories\SectionFactory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use LaraZeus\Bolt\Database\Factories\SectionFactory;
 use Spatie\Translatable\HasTranslations;
 
 /**
@@ -27,7 +28,28 @@ class Section extends Model
 
     protected $guarded = [];
 
-    protected static function newFactory(): SectionFactory
+    protected static function booted(): void
+    {
+        static::deleting(function (Section $section) {
+            if ($section->isForceDeleting()) {
+                $section->fields()->withTrashed()->get()->each(function ($item) {
+                    $item->fieldResponses()->withTrashed()->get()->each(function ($item) {
+                        $item->forceDelete();
+                    });
+                    $item->forceDelete();
+                });
+            } else {
+                $section->fields->each(function ($item) {
+                    $item->fieldResponses->each(function ($item) {
+                        $item->delete();
+                    });
+                    $item->delete();
+                });
+            }
+        });
+    }
+
+    protected static function newFactory(): Factory
     {
         return SectionFactory::new();
     }